requirements.txt
anthropic>=0.40.0
aiohttp>=3.9.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=1.5.0
jinja2>=3.1.0
websockets>=12.0
